+++ src/test/java/sonia/scm/webhook/WebHookModuleTest.java
@@ -21,35 +21,46 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-
 package sonia.scm.webhook;
 
-//~--- non-JDK imports --------------------------------------------------------
-
 import com.google.inject.AbstractModule;
-
-import org.mapstruct.factory.Mappers;
-import sonia.scm.plugin.Extension;
+import com.google.inject.Guice;
+import com.google.inject.Injector;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+import sonia.scm.net.ahc.AdvancedHttpClient;
+import sonia.scm.repository.RepositoryManager;
+import sonia.scm.store.ConfigurationStoreFactory;
 import sonia.scm.webhook.impl.AhcWebHookHttpClient;
-import sonia.scm.webhook.internal.WebHookMapper;
 
-/**
- *
- * @author Sebastian Sdorra
- */
-@Extension
-public class WebHookModule extends AbstractModule
-{
+import static org.assertj.core.api.Assertions.assertThat;
 
-  /**
-   * Method description
-   *
-   */
-  @Override
-  protected void configure()
-  {
-    bind(WebHookContext.class);
-    bind(WebHookHttpClient.class).to(AhcWebHookHttpClient.class);
-    bind(WebHookMapper.class).to(Mappers.getMapper(WebHookMapper.class).getClass());
+@ExtendWith(MockitoExtension.class)
+class WebHookModuleTest {
+
+  @Mock
+  private AdvancedHttpClient ahClient;
+
+  @Mock
+  private RepositoryManager repositoryManager;
+
+  @Mock
+  private ConfigurationStoreFactory configurationStoreFactory;
+
+  @Test
+  void shouldBindAhcWebHookHttpClient() {
+    Injector injector = Guice.createInjector(new WebHookModule(), new AbstractModule() {
+      @Override
+      protected void configure() {
+        bind(AdvancedHttpClient.class).toInstance(ahClient);
+        bind(RepositoryManager.class).toInstance(repositoryManager);
+        bind(ConfigurationStoreFactory.class).toInstance(configurationStoreFactory);
+      }
+    });
+    WebHookHttpClient client = injector.getInstance(WebHookHttpClient.class);
+    assertThat(client).isInstanceOf(AhcWebHookHttpClient.class);
   }
+
 }