+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Documentation in German ([#5](https://github.com/scm-manager/scm-webhook-plugin/pull/5))
 
 ### Changed
+- Replaced `URLConnection` based http client with `AdvancedHttpClient` ([#9](https://github.com/scm-manager/scm-webhook-plugin/pull/9))
 - Use expression language parser from scm-el-plugin ([#10](https://github.com/scm-manager/scm-webhook-plugin/pull/10))
 
 ## [2.0.0] - 2020-06-04